+++ src/atompy/physics/coltrims/__init__.py
@@ -1,3 +1,15 @@
-from ._coltrims import (
+from atompy.physics.coltrims._coltrims import (
     ion_tof_linear_fit,
 )
+
+from atompy.physics.coltrims._coulomb_explode import (
+    coulomb_explode,
+    coulomb_explode_batch,
+)
+
+
+__all__ = [
+    "ion_tof_linear_fit",
+    "coulomb_explode",
+    "coulomb_explode_batch",
+]